+++ lib/core/architecture/base.view.dart
