+++ lib/core/architecture/base.viewmodel.dart
