+++ lib/core/architecture/text.extension.dart
