+++ lib/core/base.dart
@@ -1,6 +1,6 @@
-export 'package:flutter_skyway/core/architecture/base_view.dart';
-export 'package:flutter_skyway/core/architecture/base_viewmodel.dart';
+export 'package:flutter_skyway/core/architecture/base.view.dart';
+export 'package:flutter_skyway/core/architecture/base.viewmodel.dart';
 export 'package:flutter_skyway/core/styles/text_style.dart';
-export 'package:flutter_skyway/core/architecture/text_extension.dart';
+export 'package:flutter_skyway/core/architecture/text.extension.dart';
 export 'package:flutter_skyway/generated/fonts.gen.dart';
 export 'package:flutter_skyway/generated/assets.gen.dart';
