+++ lib/data/api/data_source/news.ds.dart
@@ -1,11 +1,11 @@
 import 'package:dio/dio.dart' hide Headers;
-import 'package:flutter_skyway/data/api/api_services.dart';
-import 'package:flutter_skyway/data/api/output/users_output.dart';
-import 'package:flutter_skyway/domain/models/user.dart';
+import 'package:flutter_skyway/data/api/api.services.dart';
+import 'package:flutter_skyway/data/api/output/users.output.dart';
+import 'package:flutter_skyway/domain/entities/user.dart';
 import 'package:get/get.dart';
 import 'package:retrofit/retrofit.dart';
 
-part 'news_data_source.g.dart';
+part 'news.ds.g.dart';
 
 @RestApi(baseUrl: "https://dummyapi.io/data/v1")
 abstract class UsersDataSource {
