+++ lib/domain/entities/skyway_peer.dart
