+++ lib/domain/entities/user.dart
