+++ lib/domain/use_case/user.uc.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_skyway/core/architecture/result.dart';
-import 'package:flutter_skyway/data/api/news_data_source.dart';
-import 'package:flutter_skyway/data/repositories/user_repository.dart';
-import 'package:flutter_skyway/domain/models/user.dart';
+import 'package:flutter_skyway/data/repositories/user.repository.dart';
+import 'package:flutter_skyway/domain/entities/user.dart';
 import 'package:get/get.dart';
 
 class UserUseCase {
