+++ lib/mock/usecase/home.suc.mock.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_skyway/core/architecture/result.dart';
-import 'package:flutter_skyway/domain/models/user.dart';
+import 'package:flutter_skyway/domain/entities/user.dart';
 import 'package:flutter_skyway/presentation/home/home.suc.dart';
 
 class HomeSceneUseCaseMock implements HomeSceneUseCaseType {
