+++ lib/presentation/app/app.viewmodel.dart
@@ -1,6 +1,4 @@
 import 'package:flutter_skyway/core/base.dart';
-import 'package:flutter_skyway/data/api/news_data_source.dart';
-import 'package:flutter_skyway/data/repositories/user_repository.dart';
 import 'package:flutter_skyway/presentation/app/app.pages.dart';
 import 'package:get/get.dart';
 import 'package:mobx/mobx.dart';
